Finished the initial version of the nci gui. It does extremly poor compression and clearly has some bugs to be fixed still.
git-svn-id: svn://svn.syncleus.com/dANN/trunk@127 6ae8b97b-f314-0410-8212-aecf10b92ded

+++ projects/java_dann_examples/src/com/syncleus/core/dann/examples/nci/ui/BrainThread.java
@@ -10,7 +10,7 @@ public class BrainThread extends Thread
     private boolean busy = true;
     private BufferedImage originalImage = null;
     private BufferedImage finalImage = null;
-    private int[] compressedImage = null;
+    private byte[] compressedImage = null;
     private int width = 0;
     private int height = 0;
     private double compression;
@@ -65,7 +65,7 @@ public class BrainThread extends Thread
 
 
 
-    public int[] getLastCompressedImage()
+    public byte[] getLastCompressedImage()
     {
         return this.compressedImage;
     }
